Add mixed precision

+++ model/model_classifer.py
@@ -6,6 +6,7 @@ from keras.engine.topology import Layer
 from tensorflow.keras.initializers import Constant
 import tensorflow.keras.backend as K
 from tensorflow.keras import initializers
+import tensorflow as tf
 
 # l2 normalize
 class Normalize(Layer):
@@ -98,6 +99,6 @@ def create_classifier(source_layers, num_priors, normalizations, num_classes=21,
     # mbox_conf = Activation('softmax', name='mbox_conf_final')(mbox_conf)
 
     # predictions/concat:0, shape=(Batch, 8732, 25)
-    predictions = Concatenate(axis=2, name='predictions')([mbox_loc, mbox_conf])
+    predictions = Concatenate(axis=2, name='predictions', dtype=tf.float32)([mbox_loc, mbox_conf])
 
     return predictions
